Allow pins override of *_TIMER_NUM and HAL_*_TIMER_ISR (#18128)
