+++ Marlin/src/HAL/DUE/HAL.h
@@ -35,7 +35,6 @@
 #include "../shared/HAL_SPI.h"
 #include "fastio.h"
 #include "watchdog.h"
-#include "timers.h"
 
 #include <stdint.h>
