+++ Marlin/src/HAL/DUE/Tone.cpp
@@ -31,7 +31,6 @@
 
 #include "../../inc/MarlinConfig.h"
 #include "HAL.h"
-#include "timers.h"
 
 static pin_t tone_pin;
 volatile static int32_t toggles;
