+++ Marlin/src/HAL/DUE/timers.cpp
@@ -34,8 +34,6 @@
 #include "../../inc/MarlinConfig.h"
 #include "HAL.h"
 
-#include "timers.h"
-
 // ------------------------
 // Local defines
 // ------------------------
