+++ Marlin/src/HAL/DUE/timers.h
@@ -40,11 +40,17 @@ typedef uint32_t hal_timer_t;
 #define HAL_TIMER_RATE         ((F_CPU) / 2)    // frequency of timers peripherals
 
 #ifndef STEP_TIMER_NUM
-#define STEP_TIMER_NUM 2  // index of timer to use for stepper
+  #define STEP_TIMER_NUM        2  // Timer Index for Stepper
 #endif
-#define TEMP_TIMER_NUM 4  // index of timer to use for temperature
-#define PULSE_TIMER_NUM STEP_TIMER_NUM
-#define TONE_TIMER_NUM 6  // index of timer to use for beeper tones
+#ifndef PULSE_TIMER_NUM
+  #define PULSE_TIMER_NUM       STEP_TIMER_NUM
+#endif
+#ifndef TEMP_TIMER_NUM
+  #define TEMP_TIMER_NUM        4  // Timer Index for Temperature
+#endif
+#ifndef TONE_TIMER_NUM
+  #define TONE_TIMER_NUM        6  // index of timer to use for beeper tones
+#endif
 
 #define TEMP_TIMER_FREQUENCY   1000 // temperature interrupt frequency
 
@@ -66,8 +72,12 @@ typedef uint32_t hal_timer_t;
 #ifndef HAL_STEP_TIMER_ISR
   #define HAL_STEP_TIMER_ISR() void TC2_Handler()
 #endif
-#define HAL_TEMP_TIMER_ISR()  void TC4_Handler()
-#define HAL_TONE_TIMER_ISR()  void TC6_Handler()
+#ifndef HAL_TEMP_TIMER_ISR
+  #define HAL_TEMP_TIMER_ISR() void TC4_Handler()
+#endif
+#ifndef HAL_TONE_TIMER_ISR
+  #define HAL_TONE_TIMER_ISR() void TC6_Handler()
+#endif
 
 // ------------------------
 // Types
