+++ Marlin/src/HAL/ESP32/HAL.cpp
@@ -21,14 +21,12 @@
  */
 #ifdef ARDUINO_ARCH_ESP32
 
-#include "HAL.h"
-#include "timers.h"
+#include "../../inc/MarlinConfig.h"
+
 #include <rom/rtc.h>
 #include <driver/adc.h>
 #include <esp_adc_cal.h>
 #include <HardwareSerial.h>
-
-#include "../../inc/MarlinConfigPre.h"
 
 #if ENABLED(WIFISUPPORT)
   #include <ESPAsyncWebServer.h>
