+++ Marlin/src/HAL/ESP32/HAL.h
@@ -34,8 +34,6 @@
 #include "watchdog.h"
 #include "i2s.h"
 
-#include "timers.h"
-
 #if ENABLED(WIFISUPPORT)
   #include "WebSocketSerial.h"
 #endif
