+++ Marlin/src/HAL/ESP32/timers.cpp
@@ -27,9 +27,7 @@
 #include <driver/periph_ctrl.h>
 #include <driver/timer.h>
 
-#include "HAL.h"
-
-#include "timers.h"
+#include "../../inc/MarlinConfig.h"
 
 // ------------------------
 // Local defines
