+++ Marlin/src/HAL/LINUX/HAL.h
@@ -56,7 +56,6 @@ uint8_t _getc();
 #include "../shared/HAL_SPI.h"
 #include "fastio.h"
 #include "watchdog.h"
-#include "timers.h"
 #include "serial.h"
 
 #define SHARED_SERVOS HAS_SERVOS
