+++ Marlin/src/HAL/LINUX/timers.cpp
@@ -24,7 +24,6 @@
 #include "hardware/Timer.h"
 
 #include "../../inc/MarlinConfig.h"
-#include "timers.h"
 
 /**
  * Use POSIX signals to attempt to emulate Interrupts
