+++ Marlin/src/HAL/LPC1768/timers.cpp
@@ -29,7 +29,6 @@
 #ifdef TARGET_LPC1768
 
 #include "../../inc/MarlinConfig.h"
-#include "timers.h"
 
 void HAL_timer_init() {
   SBI(LPC_SC->PCONP, SBIT_TIMER0);  // Power ON Timer 0
