+++ Marlin/src/HAL/SAMD51/timers.cpp
@@ -24,7 +24,6 @@
 // Includes
 // --------------------------------------------------------------------------
 #include "../../inc/MarlinConfig.h"
-#include "timers.h"
 
 // --------------------------------------------------------------------------
 // Local defines
