+++ Marlin/src/HAL/SAMD51/timers.h
@@ -32,9 +32,15 @@ typedef uint32_t hal_timer_t;
 
 #define HAL_TIMER_RATE      F_CPU   // frequency of timers peripherals
 
-#define STEP_TIMER_NUM      0  // index of timer to use for stepper (also +1 for 32bits counter)
-#define PULSE_TIMER_NUM     STEP_TIMER_NUM
-#define TEMP_TIMER_NUM      RTC_TIMER_NUM  // index of timer to use for temperature
+#ifndef STEP_TIMER_NUM
+  #define STEP_TIMER_NUM        0  // Timer Index for Stepper
+#endif
+#ifndef PULSE_TIMER_NUM
+  #define PULSE_TIMER_NUM       STEP_TIMER_NUM
+#endif
+#ifndef TEMP_TIMER_NUM
+  #define TEMP_TIMER_NUM        RTC_TIMER_NUM // Timer Index for Temperature
+#endif
 
 #define TEMP_TIMER_FREQUENCY   1000 // temperature interrupt frequency
 
@@ -57,16 +63,18 @@ typedef uint32_t hal_timer_t;
                                : (t == TEMP_TIMER_NUM) ? 6                        \
                                : 7
 
-#define _TC_HANDLER(t)            void TC##t##_Handler()
-#define TC_HANDLER(t)             _TC_HANDLER(t)
-#define HAL_STEP_TIMER_ISR()      TC_HANDLER(STEP_TIMER_NUM)
+#define _TC_HANDLER(t)          void TC##t##_Handler()
+#define TC_HANDLER(t)           _TC_HANDLER(t)
+#ifndef HAL_STEP_TIMER_ISR
+  #define HAL_STEP_TIMER_ISR()  TC_HANDLER(STEP_TIMER_NUM)
+#endif
 #if STEP_TIMER_NUM != PULSE_TIMER_NUM
-  #define HAL_PULSE_TIMER_ISR()   TC_HANDLER(PULSE_TIMER_NUM)
+  #define HAL_PULSE_TIMER_ISR() TC_HANDLER(PULSE_TIMER_NUM)
 #endif
 #if TEMP_TIMER_NUM == RTC_TIMER_NUM
-  #define HAL_TEMP_TIMER_ISR()    void RTC_Handler()
+  #define HAL_TEMP_TIMER_ISR()  void RTC_Handler()
 #else
-  #define HAL_TEMP_TIMER_ISR()    TC_HANDLER(TEMP_TIMER_NUM)
+  #define HAL_TEMP_TIMER_ISR()  TC_HANDLER(TEMP_TIMER_NUM)
 #endif
 
 // --------------------------------------------------------------------------
