+++ Marlin/src/HAL/STM32/timers.cpp
@@ -21,8 +21,6 @@
  */
 #if defined(ARDUINO_ARCH_STM32) && !defined(STM32GENERIC)
 
-#include "timers.h"
-
 #include "../../inc/MarlinConfig.h"
 
 // ------------------------
