+++ Marlin/src/HAL/STM32F1/HAL.h
@@ -36,7 +36,6 @@
 #include "fastio.h"
 #include "watchdog.h"
 
-#include "timers.h"
 
 #include <stdint.h>
 #include <util/atomic.h>
