+++ Marlin/src/HAL/STM32F1/Servo.cpp
@@ -29,7 +29,6 @@
 uint8_t ServoCount = 0;
 
 #include "Servo.h"
-#include "timers.h"
 
 //#include "Servo.h"
