+++ Marlin/src/HAL/STM32F1/timers.cpp
@@ -27,8 +27,6 @@
 #ifdef __STM32F1__
 
 #include "../../inc/MarlinConfig.h"
-#include "HAL.h"
-#include "timers.h"
 
 // ------------------------
 // Local defines
