+++ Marlin/src/HAL/STM32_F4_F7/HAL.h
@@ -31,7 +31,6 @@
 #include "../shared/HAL_SPI.h"
 
 #include "fastio.h"
-#include "timers.h"
 #include "watchdog.h"
 
 #include <stdint.h>
