+++ Marlin/src/HAL/STM32_F4_F7/STM32F7/timers.cpp
@@ -21,8 +21,7 @@
  */
 #if defined(STM32GENERIC) && defined(STM32F7)
 
-#include "../HAL.h"
-#include "timers.h"
+#include "../../../inc/MarlinConfig.h"
 
 // ------------------------
 // Local defines
