+++ Marlin/src/HAL/TEENSY31_32/HAL.h
@@ -34,7 +34,6 @@
 #include "fastio.h"
 #include "watchdog.h"
 
-#include "timers.h"
 
 #include <stdint.h>
