+++ Marlin/src/HAL/TEENSY35_36/HAL.h
@@ -34,8 +34,6 @@
 #include "fastio.h"
 #include "watchdog.h"
 
-#include "timers.h"
-
 #include <stdint.h>
 #include <util/atomic.h>
