+++ Marlin/src/HAL/TEENSY35_36/timers.cpp
@@ -27,8 +27,7 @@
 
 #if defined(__MK64FX512__) || defined(__MK66FX1M0__)
 
-#include "HAL.h"
-#include "timers.h"
+#include "../../inc/MarlinConfig.h"
 
 /** \brief Instruction Synchronization Barrier
   Instruction Synchronization Barrier flushes the pipeline in the processor,
