+++ Marlin/src/inc/MarlinConfig.h
@@ -30,6 +30,7 @@
 #include "../HAL/HAL.h"
 
 #include "../pins/pins.h"
+#include HAL_PATH(../HAL, timers.h)
 #include HAL_PATH(../HAL, spi_pins.h)
 
 #include "Conditionals_post.h"
