+++ Marlin/src/pins/sam/pins_ARCHIM1.h
@@ -46,6 +46,7 @@
 //
 // Timers
 //
+// These are already defined in DUE, so must be undefined first
 #define STEP_TIMER_NUM                         3
 #define HAL_STEP_TIMER_ISR()  void TC3_Handler()
